pyproject.toml
[tool.poetry]
name = "incolume-py-model20230304"
version = "0.2.0"
description = ""
authors = ["Britodfbr <britodfbr@gmail.com>"]
readme = "README.md"
packages = [{include = "incolume"}]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
blue = "^0.9.1"
isort = "^5.12.0"
mypy = "^1.0.1"
taskipy = "^1.10.3"

[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.0"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
#log_format = '%(asctime)s %(levelname)s %(message)s'
log_format = '%(asctime)s; %(levelname)-8s; %(name)s; %(module)s; %(funcName)s; %(message)s'
log_date_format ='%FT%T%z'
addopts = "--cov=incolume"
testpaths = [
  "tests",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "webtest: marks tests as webtest (deselect with '-m \"not webtest\"')",
  "fasttest: marks tests as fasttest (deselect with '-m \"not fasttest\"')",
]

[tool.coverage.html]
directory = "coverage_report/html"

[tool.coverage.run]
omit = [
  # omit anything in a .local directory anywhere
  "*/.local/*",
  # omit everything in /usr
  "/usr/*",
  # omit this single file
  "configure/truncus.py",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "__package__ =",
  "__author__",
  "def run",
]
